pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Pineapple_Juice <st.okopnyi@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
validators = "^0.20.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
